feat(text_layers.py): Assign blank text to text item if that is what is given to the text formatting class
Text fields in the PSD files usually have placeholder text in them, so in some rare situations, e.g. with Tokens, where the text should be cleared it isn't unless that specific case is separately handled in the template code. This change allows relying on the default text processing logic in such situations.